(feat) move base files and Bamboo Relay to new location
