+++ hummingbot/connector/exchange/bamboo_relay/bamboo_relay_constants.py
@@ -35,4 +35,4 @@ ZERO_EX_TEST_COORDINATOR_ADDRESS = "0x4d3d5c850dd5bd9d6f4adda3dd039a3c8054ca29"
 ZERO_EX_TEST_COORDINATOR_REGISTRY_ADDRESS = "0xaa86dda78e9434aca114b6676fc742a18d15a1cc"
 BAMBOO_RELAY_TEST_FEE_RECIPIENT_ADDRESS = "0x5409ed021d9299bf6814279a6a1411a7e866a631"
 
-ZERO_EX_PROTOCOL_FEE_MULTIPLIER = 150000 # Full fee = 150,000 * gasPrice * orderCount
+ZERO_EX_PROTOCOL_FEE_MULTIPLIER = 150000  # Full fee = 150,000 * gasPrice * orderCount
