+++ hummingbot/connector/exchange/bamboo_relay/bamboo_relay_market.pxd
