+++ hummingbot/connector/exchange/bamboo_relay/bamboo_relay_order_book_message.py
