+++ hummingbot/market/paper_trade/__init__.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from hummingbot.market.bamboo_relay.bamboo_relay_order_book_tracker import BambooRelayOrderBookTracker
+from hummingbot.connector.exchange.bamboo_relay.bamboo_relay_order_book_tracker import BambooRelayOrderBookTracker
 from hummingbot.connector.exchange.binance.binance_order_book_tracker import BinanceOrderBookTracker
 from hummingbot.market.coinbase_pro.coinbase_pro_order_book_tracker import CoinbaseProOrderBookTracker
 from hummingbot.market.huobi.huobi_market import HuobiMarket
@@ -12,7 +12,7 @@ from hummingbot.market.radar_relay.radar_relay_order_book_tracker import RadarRe
 from hummingbot.market.dolomite.dolomite_order_book_tracker import DolomiteOrderBookTracker
 from hummingbot.market.dolomite.dolomite_market import DolomiteMarket
 
-from hummingbot.market.bamboo_relay.bamboo_relay_market import BambooRelayMarket
+from hummingbot.connector.exchange.bamboo_relay.bamboo_relay_market import BambooRelayMarket
 from hummingbot.connector.exchange.binance.binance_market import BinanceMarket
 from hummingbot.market.coinbase_pro.coinbase_pro_market import CoinbaseProMarket
